fix: data loading issues with htfa notebook
htfa notebook expects the dataset to be in /data. This is where it
resides in the Docker image. I extracted the download commands for
the data files and put them into download_data.sh. I modified the
commands to extract the data into a local data/ folder. I then
modified the notebook to look for the data in both of these locations
and throw an exception otherwise.